+++ CHANGELOG.md
@@ -2,10 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.3]
+
+### ADD Listner for prediction loading
+
+- Implement a nullable listner to show loading status of predictions.
+
 ## [0.0.2]
-### Initial Release
+
+### Update README
+
 - Update README file.
 
 ## [0.0.1]
+
 ### Initial Release
+
 - First release of `google_places_autocomplete`, providing basic functionality.